backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.25
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
websockets==12.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
aiosqlite==0.19.0
pydantic-settings==2.1.0
gridstatusio==0.13.0
kafka-python==2.0.2
